Use page actions and declarative content rules
Switch from browser actions to page actions, and use declarative content
rules to do away with the content script altogether. This doesn't give
the required functionality though since injecting a script via
declarative content (RequestContentScript) is still experimental (?),
and activeTab permissions (that allow us to inject a script on the
active tab) are enabled by the user clicking on the page action ican.

So, dead end, but left here for reference.